+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Make `get_name` and `get_bcid` methods available for all BaseEntity objects.
 - provide a selection system for the type of name to use for the objects.
-
+- Add parameter in MemoryBase get_data() function to change the label_type ("name" or "bcid") 
+- Add parameter in MemoryBase get_data() function to return a pandas DataFrame
 
 ### Fixed
  - Fix the parsing of date list containing "null" elements (https://gitlab.ipleanware.com/braincube/misc/redmine/-/issues/2597)